Factor out _write() implementation
