+++ utils/CMakeLists.txt
@@ -5,6 +5,7 @@
 if(BUILD_WITH_CMSIS)
   if(ARM_TARGET STREQUAL "NRF52840")
     set(UTILS_SRC ${CMAKE_CURRENT_SOURCE_DIR}/nrf52840_cmsis.c)
+    set(UTILS_SRC "${UTILS_SRC};${CMAKE_CURRENT_SOURCE_DIR}/nrf_write_cmsis.c")
   elseif(ARM_TARGET MATCHES "^STM32F4")
     set(UTILS_SRC ${CMAKE_CURRENT_SOURCE_DIR}/stm32f4_cmsis.c)
   elseif(ARM_TARGET MATCHES "^STM32F7")
@@ -13,6 +14,9 @@ if(BUILD_WITH_CMSIS)
     set(UTILS_SRC ${CMAKE_CURRENT_SOURCE_DIR}/stm32l4_cmsis.c)
   elseif(ARM_TARGET STREQUAL "CORSTONE-300")
     set(UTILS_SRC ${CMAKE_CURRENT_SOURCE_DIR}/corstone-300.c)
+  endif()
+  if(ARM_TARGET MATCHES "^STM32")
+    set(UTILS_SRC "${UTILS_SRC};${CMAKE_CURRENT_SOURCE_DIR}/stm32_write_cmsis.c")
   endif()
 elseif(BUILD_WITH_LIBOPENCM3)
   set(UTILS_SRC ${CMAKE_CURRENT_SOURCE_DIR}/stm32f4_libopencm3.c)
