+++ utils/nrf52840_cmsis.c
@@ -5,7 +5,6 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include <errno.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <unistd.h>
@@ -42,22 +41,6 @@ void send_blocking(uint16_t c) {
 #ifdef USE_UART0
   nrfx_uart_tx(&uart, (uint8_t*)&c, 1);
 #endif
-}
-
-int _write(int file, char* ptr, int len) {
-  int i;
-
-  if (file == STDOUT_FILENO || file == STDERR_FILENO) {
-    for (i = 0; i < len; i++) {
-      if (ptr[i] == '\n') {
-        send_blocking('\r');
-      }
-      send_blocking(ptr[i]);
-    }
-    return i;
-  }
-  errno = EIO;
-  return -1;
 }
 
 bool wait_until(uint64_t nanos) { return true; }
