+++ utils/stm32f4_cmsis.c
@@ -5,7 +5,6 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include <errno.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <unistd.h>
@@ -214,22 +213,6 @@ void usart_send_blocking(uint8_t c) {
   while (!(USART2->SR & USART_SR_TXE))
     ;
 #endif
-}
-
-int _write(int file, char* ptr, int len) {
-  int i;
-
-  if (file == STDOUT_FILENO || file == STDERR_FILENO) {
-    for (i = 0; i < len; i++) {
-      if (ptr[i] == '\n') {
-        usart_send_blocking('\r');
-      }
-      usart_send_blocking(ptr[i]);
-    }
-    return i;
-  }
-  errno = EIO;
-  return -1;
 }
 
 bool wait_until(uint64_t nanos) { return true; }
